:truck: chore(api) move student info for accoun recovery serializer
