+++ api/src/identity-access-management/infrastructure/serializers/jsonapi/student-information-for-account-recovery-serializer.js
