+++ api/tests/identity-access-management/unit/infrastructure/serializers/jsonapi/student-information-for-account-recovery-serializer_test.js
@@ -1,6 +1,6 @@
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/student-information-for-account-recovery-serializer.js';
-import { StudentInformationForAccountRecovery } from '../../../../../src/shared/domain/read-models/StudentInformationForAccountRecovery.js';
-import { expect } from '../../../../test-helper.js';
+import * as serializer from '../../../../../../src/identity-access-management/infrastructure/serializers/jsonapi/student-information-for-account-recovery-serializer.js';
+import { StudentInformationForAccountRecovery } from '../../../../../../src/shared/domain/read-models/StudentInformationForAccountRecovery.js';
+import { expect } from '../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | student-information-for-account-recovery-serializer', function () {
   describe('#serialize()', function () {
